# tsdb/index: fix BenchmarkIntersect to do work on each loop