Previously all the postings constructed were consumed on the first iteration, so subsequent iterations did no work.

After fixing this, we see that `ExpandPostings` is causing most of the memory allocations on each iteration, so stop doing that.

Before (stress these are meaningless):
```
BenchmarkIntersect/LongPostings1-8              22208444                49.40 ns/op           96 B/op          2 allocs/op
BenchmarkIntersect/LongPostings2-8              20523910                51.64 ns/op           99 B/op          2 allocs/op
BenchmarkIntersect/ManyPostings-8                   5889            206882 ns/op              32 B/op          1 allocs/op
```

After:
```
BenchmarkIntersect/LongPostings1-8                 35192             34243 ns/op             224 B/op          6 allocs/op
BenchmarkIntersect/LongPostings2-8                    25          46704775 ns/op             224 B/op          6 allocs/op
BenchmarkIntersect/ManyPostings-8                      7         150956738 ns/op              32 B/op          1 allocs/op
```